+++ package.json
@@ -52,5 +52,10 @@
     "ts-node": "7.0.1",
     "tslint": "5.11.0",
     "typescript": "3.0.1"
+  },
+  "prettier": {
+    "semi": false,
+    "trailingComma": "all",
+    "singleQuote": true
   }
 }
