+++ src/generators/reason-generator.ts
@@ -4,7 +4,7 @@ import * as camelCase from "camelcase";
 import * as refmt from "reason";
 import { GraphQLScalarType, GraphQLTypeField } from "../source-helper";
 
-import { GenerateArgs } from "./generator-interface";
+import { GenerateArgs } from "./types";
 
 type SpecificGraphQLScalarType =
   | "bool"
